packages/backend/requirements.txt
fastapi==0.104.1
pydantic==2.5.2
PyJWT==2.8.0
PyJWT==2.8.0
uvicorn==0.24.0.post1
